+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 * Added several tests
 * Added TPMS status (`vehicle.tires.system_status`)
 * Added GPS coordinates (`vehicle.position`)
+* Added oil life remaining (`vehicle.oil_life_remaining`)
+* Added seatbelt status (`vehicle.seatbelts`)
 * Refactored tires to include TPMS status for each tire (`vehicle.tires.front_right.status`)
 
 # v0.3.0